Switch page padding from sections to body
Instead of applying the padding to each section of the page, I switched
it so that this padding is only applied once on the body. I also gave
the body a unique background color to distiniguish it from the website
content.

I'm unsure if this looks good from a design perspective, but it will at
least help me to better organize the contents of the page.