Reverted forced aspect ratio
I realized I made a mistake in taking the design reference image too
literally. I assumed I had to fit everything I saw in the image onto
exactly one page. Realistically, I can just use a vertical scroll to go
through all of the content on the page since it is a website after all.

Now I will just set a min-height in the body and let it expand downward
as more content is added to the page.