Create one selector for all page sections
I noticed that the different sections of the page all have some
properties in common, such as the background-color expanding to fill
the full width of the page, and having border-box on.

Additionally this preps me to add a consistent padding on the left and
right sides of the page.