Change name of ad-block to action-block
My browser ad-blocker was removing ad-block from the page because it had
"ad" in the title. I had mistakenly named it ad-block at the the
beginning because it looked like one of those banner ads but looking
closer it's part of the website itself, so this name change is more
fitting.